refactor: split `Chats` into different files
