+++ frontend/src/components/Overlay.tsx
@@ -1,0 +1,3 @@
+export const Overlay = () => {
+  return <div className="fixed inset-0 z-[1000] bg-gray-900/50"></div>;
+};
