+++ frontend/src/components/index.ts
@@ -1,0 +1,15 @@
+export * from "./AddFriendButton";
+export * from "./Alert";
+export * from "./Avatar";
+export * from "./Badge";
+export * from "./Button";
+export * from "./Card";
+export * from "./Hr";
+export * from "./Input";
+export * from "./Overlay";
+export * from "./Paginate";
+export * from "./Select";
+export * from "./Table";
+export * from "./Typography";
+export * from "./UserForm";
+export * from "./UserUpdateAvatar";
