+++ frontend/src/components/nav/Sidebar.tsx
@@ -12,7 +12,7 @@ import { INavitem } from "../../@interfaces";
 import classNames from "classnames";
 import { LuLogOut } from "react-icons/lu";
 import { Navitem } from "./Navitem";
-import { User } from "../../types/types";
+import { User } from "../../types";
 import { Avatar } from "../Avatar";
 import { FaArrowLeft, FaArrowRight } from "react-icons/fa";
 
@@ -88,7 +88,7 @@ export const Sidebar: FC<ISidebarProps> = ({
           {
             block: isOpen,
             hidden: !isOpen,
-          },
+          }
         )}
       ></div>
 
@@ -102,7 +102,7 @@ export const Sidebar: FC<ISidebarProps> = ({
           "shadow-xl",
           "z-[999] w-[16rem] max-w-[16rem] h-full top-0 left-0 md:relative fixed",
           "border-r border-gray-100 dark:border-opacity-10",
-          className,
+          className
         )}
       >
         <div
@@ -110,7 +110,7 @@ export const Sidebar: FC<ISidebarProps> = ({
             "flex ml-3 py-2 mt-5 mb-4 transform ease-out duration-100",
             {
               "px-5": isOpen,
-            },
+            }
           )}
         >
           <img
@@ -127,7 +127,7 @@ export const Sidebar: FC<ISidebarProps> = ({
               {
                 "opacity-100 delay-150": isOpen,
                 "opacity-0 delay-0": !isOpen,
-              },
+              }
             )}
             customWeight="medium"
           >
@@ -141,7 +141,7 @@ export const Sidebar: FC<ISidebarProps> = ({
               className={classNames(
                 "absolute top-0 flex items-center justify-center",
                 "w-8 h-8",
-                "bg-gray-800 rounded-full cursor-pointer -right-3",
+                "bg-gray-800 rounded-full cursor-pointer -right-3"
               )}
               onClick={() => !isTab && setIsOpen(!isOpen)}
             >
@@ -160,7 +160,7 @@ export const Sidebar: FC<ISidebarProps> = ({
               "whitespace-nowrap px-2.5 py-5 overflow-x-hidden no-scrollbar",
               {
                 "h-[60%]": isMob,
-              },
+              }
             )}
           >
             {options.map((item) => (
